@@ -12,18 +12,20 @@
   },
   "dependencies": {
     "@element-plus/icons-vue": "^2.3.1",
+    "@vicons/ionicons5": "^0.12.0",
     "@vueuse/core": "^11.1.0",
     "axios": "^1.7.7",
     "core-js": "^3.8.3",
-    "element-plus": "^2.8.4",
     "moment": "^2.30.1",
-    "vue": "^3.2.13",
+    "naive-ui": "^2.40.1",
+    "vue": "^3.5.12",
     "vue-router": "^4.4.5",
     "vuex": "^4.1.0"
   },
   "devDependencies": {
-    "@babel/core": "^7.12.16",
-    "@babel/eslint-parser": "^7.12.16",
+    "@babel/core": "^7.25.8",
+    "@babel/eslint-parser": "^7.25.8",
+    "@eslint/js": "^9.13.0",
     "@vue/cli-plugin-babel": "~5.0.0",
     "@vue/cli-plugin-eslint": "~5.0.0",
     "@vue/cli-service": "~5.0.0",
@@ -31,13 +33,15 @@
     "css-loader": "^7.1.2",
     "eslint": "^8.57.1",
     "eslint-config-prettier": "^9.1.0",
-    "eslint-plugin-vue": "^9.28.0",
+    "eslint-plugin-vue": "^9.29.1",
+    "eslint-webpack-plugin": "^4.2.0",
+    "globals": "^15.11.0",
     "postcss": "^8.4.47",
     "prettier": "^3.3.3",
-    "sass": "^1.79.4",
+    "sass": "^1.80.3",
     "sass-loader": "^16.0.2",
     "style-loader": "^4.0.0",
-    "tailwindcss": "^3.4.13",
+    "tailwindcss": "^3.4.14",
     "webpack-bundle-analyzer": "^4.10.2"
   },
   "eslintConfig": {
