@@ -1,4 +1,4 @@
-import { useDark, useToggle } from "@vueuse/core";
+import { useDark, useToggle } from '@vueuse/core';
 
 export const isDark = useDark();
 export const toggleDark = useToggle(isDark);
